refactor file structure for stability
