+++ miner_objects/order_watcher/modules/ccxt/isFilled.js
@@ -1,0 +1,5 @@
+export function isFilled(order) {
+  return (
+    order?.info?.orderStatus === "Filled" && order?.info?.state === "filled"
+  );
+}
