+++ miner_objects/order_watcher/modules/order/toSignal.js
@@ -1,0 +1,7 @@
+export function toSignal(order, leverage = 0) {
+  return {
+    trade_pair: order.info.symbol,
+    order_type: order.side.toUpperCase() === "BUY" ? "LONG" : "SHORT",
+    leverage: leverage,
+  };
+}
