+++ miner_objects/order_watcher/modules/websocket/broadcast.js
@@ -1,13 +1,3 @@
-import { WebSocket, WebSocketServer } from "ws";
-
-let wss = null;
-
-export function initializeWebSocketServer(server) {
-  wss = new WebSocketServer({ server });
-  console.log("WebSocket server initialized");
-  return wss;
-}
-
 export function broadcast(wss, message) {
   if (!wss) {
     console.error("WebSocket server is not initialized.");
