+++ miner_objects/order_watcher/modules/websocket/start.js
@@ -1,0 +1,9 @@
+import { WebSocketServer } from "ws";
+
+let wss = null;
+
+export function startWebSocket(server) {
+  wss = new WebSocketServer({ server });
+  console.log("WebSocket server initialized");
+  return wss;
+}
